backend/requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12

# ── Google Gemini ────────────────────────────────────────────────────────────
google-generativeai==0.8.0

# ── Vector Store ─────────────────────────────────────────────────────────────
faiss-cpu==1.8.0
numpy==1.26.4

# ── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.27.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.0
pytest-asyncio==0.23.0
httpx  # for TestClient
